feat(dns): extract addrinfo conversion utility from HappyEyeballs
Extract he_convert_resolver_result() and he_free_converted_addrinfo()
from SocketHappyEyeballs.c to SocketCommon as shared utilities.

Changes:
- Add SocketCommon_resolver_to_addrinfo() - Convert SocketDNSResolver_Result to addrinfo
- Add SocketCommon_free_resolver_addrinfo() - Free converted addrinfo chain
- Update SocketHappyEyeballs.c to use shared functions
- Use void* in header to avoid circular dependency on DNS resolver types
- Maintain embedded sockaddr allocation pattern for single-free cleanup

This enables code reuse across DNS-aware socket modules while maintaining
the existing allocation semantics (sockaddr embedded in addrinfo block).

Closes #1054
